fix(endgame): faithful forge/wheel/prey/quickloot fixes + crash resilience
Forge:
- read all tunables from config.lua (costs/rates/limits) instead of
  hardcoding; build the history entry description (0x88 was blank).

Wheel of Destiny:
- canUse gated on vocation + level>50; getOptions honors temple PZ
  (can decrease only there); vocation byte handles Monk.

Prey / Task Hunter:
- prey reroll price and price bytes read from config.lua; prey bonus %
  by type (Damage 2r+5, Defense 2r+10, XP/Loot 3r+10); task reward via
  the TaskHuntingOption table (TargetKills + claim boost math) instead of
  hardcoded 200/flat-10; fix stale opcode comment.

Quick Loot:
- resolve managed containers by inventory container id (loot containers
  live in the inventory, not on map tiles); honor fallback-to-main and
  the DEFAULT-category fallback; variant 0/1/2 semantics with
  QUICK_LOOT_MAX_CORPSES cap; report container ids in 0xC0.

Crash resilience / teleports:
- recover() around the per-packet handler so a scripting error or nil
  deref (e.g. a boss teleport) no longer takes down the whole server.
- add Player:canFightBoss / setBossCooldown / getBossCooldown bindings
  (used by boss entrance scripts like Drume).
- Tile:hasFlag reads its arg leniently (nil -> 0, like C++ getNumber) and
  the TILESTATE_* Lua enum matches the OTBM tile-flag bits actually
  stored, adding the missing constants (HOUSE, TELEPORT, DEPOT, ...) so
  scripts like the adventurers' guild stone stop erroring.

+++ internal/game/forge_test.go
@@ -88,15 +88,15 @@ func TestForgeResourceConversionDustToSlivers(t *testing.T) {
 	if p.GetForgeDusts() != 40 {
 		t.Fatalf("dusts = %d, want 40", p.GetForgeDusts())
 	}
-	if got := p.GetForgeSlivers(cat); got != ForgeSliverAmount {
-		t.Fatalf("slivers = %d, want %d", got, ForgeSliverAmount)
+	if got := p.GetForgeSlivers(cat); got != uint32(ForgeSliverAmount()) {
+		t.Fatalf("slivers = %d, want %d", got, ForgeSliverAmount())
 	}
 }
 
 func TestForgeResourceConversionSliversToCore(t *testing.T) {
 	cat := forgeCatalog()
 	p := forgePlayer()
-	addToBackpack(p, &Item{ID: ItemForgeSliver, Count: ForgeCoreCost})
+	addToBackpack(p, &Item{ID: ItemForgeSliver, Count: uint16(ForgeCoreCost())})
 
 	if !p.ForgeResourceConversion(cat, ForgeActionSliverToCore) {
 		t.Fatalf("sliver->core should succeed")
